cleaned up server, prepping for mongo
