+++ package.json
@@ -29,6 +29,7 @@
     "concurrently": "^5.2.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "mongoose": "^5.9.21",
     "ncp": "^2.0.0",
     "nodemailer": "^6.4.6",
     "socket.io": "^2.3.0"
